Fix overflow on malformed cursor files
The code did a plain "4 * width * height" with the values read from the
file without worrying about overflows. This overflow is easy to trigger.

In this commit, I add a new test for this. This test panics (in debug
builds) with "attempt to multiply with overflow". Then I fix this test
by using checked multiplication.

Signed-off-by: Uli Schlachter <psychon@znc.in>